Ahoy! (This probably has been done before but I do not care.)
In this commit I did the following:

Removed most recipes and advancements totally not because I want to increase my commit power and actually because I wanted to wipe it away as they may conflict with the new recipes.

Bumped to 0.4 but its not ready yet guys its not 0.4
Added some more sensing for a future project.
Added a tag that was needed.

So whos ready for the big one?