EnhancedRegistry (#102)
* Made a prototype registry

In this commit I did the following:
Made a prototype "Enhanced" registry.

This is a VERY early prototype, and does not have many features even added. As an extra Siuol bonus, I didnt even test the previous commit on master nor this one!

Signed-off-by: Siuol <43890828+Siuolthepic@users.noreply.github.com>

* Fully done registry

In this commit I did the following:
Enhanced the registry, its now for blocks AND items.

This is still a VERY early prototype. And a stupid idea as well, but you have to counterract the amount of effort it takes to maintain the registry somehow.

Signed-off-by: Siuol <43890828+Siuolthepic@users.noreply.github.com>

* Completed* Enhanced Registry

In this commit I did the following:
Finished up the enhanced registry for now, it contains most blocks/items. The amount of concerning code here is unreal. It still technically needs support for Mossy/Cobble, but for now, it should work well.

I am very proud of my self. Look important people I did a thing.

Signed-off-by: Siuol <43890828+Siuolthepic@users.noreply.github.com>

* No its not 1.18.2 yet guys.

In this commit I did the following:
- Updated models for Sandstone, Red Sandstone, and Soul Sandstone to look improved.
- Cleaned up EnhancedRegistry code.
- Removed BlockColors, replaced with the built-in DyeColor enum.
- Turned Wither Skele drops into a loot table injection on Fabric. Forge has its regular version until I can get to that.

1.18.2 SOON I PROMISE

Signed-off-by: Siuol <43890828+Siuolthepic@users.noreply.github.com>

* Oh cmon its close enough.

In this commit I did the following:
- Got us on 1.18.2 to a frame of the loading screen in game.

Ima be honest, I already need a break.

Signed-off-by: Siuol <43890828+Siuolthepic@users.noreply.github.com>

* HAHA YES I FINALLY DID IT

In this commit I did the following:
- Got us into the game on Fabric 1.18.2

I already want another break.

Signed-off-by: Siuol <43890828+Siuolthepic@users.noreply.github.com>

* Someone else please pick up from here

I have lost like 5 years from my life trying to make this mod work on Forge.

Signed-off-by: Siuol <43890828+Siuolthepic@users.noreply.github.com>

* Alright right on schedule yep. (0.5.0-RC2)

Screw the commit names this is 0.5.0-RC2

Signed-off-by: Siuol <43890828+Siuolthepic@users.noreply.github.com>